rag-chatbot/backend/requirements.txt
fastapi
pinecone
openai
uvicorn
python-dotenv
psycopg2-binary
supabase
PyJWT
requests
bcrypt
pydantic
fastapi-users